requirements.txt
asgiref==3.8.1
click==8.1.8
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-redis==5.4.0
django-rq==2.10.3
djangorestframework==3.15.2
redis==5.2.1
rq==1.16.2
sqlparse==0.5.3
whitenoise==6.8.2
